+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 - Fix `DropdownPhoneSelect` when phone number is empty.
 
+- Feature: Add `disableInput` prop to `PillNumberInput`, defaults to `false`.
+
 ## [2.100.1] - 2020-11-26
 
 Fix: